refactor(routes): extract settings + admin routes from server.ts (#3502 PR2)
Move the 34 remaining inline `/settings/*` (17) and `/admin/*` (17)
apiRouter handlers out of server.ts and into route modules, per Task 4.1
PR2 of the #3962 remediation epic.

- Extend routes/settingsRoutes.ts with the 17 inline /settings/* handlers
  (traceroute/remote-localstats/time-sync automation config, auto-ping,
  key-repair/distance-delete logs, position-estimation run-now/status,
  mark-all-welcomed). Bodies moved verbatim; only the leading
  apiRouter.<method>('/settings/...') registration line was rewritten to
  router.<method>('/...'). Adds resolveSourceManager,
  validateFilterNameRegexOnSave, positionEstimationScheduler, and
  autoDeleteByDistanceService imports the new handlers close over.

- Create routes/adminRoutes.ts (new module, mounted at '/admin', all
  handlers requireAdmin()) with the 17 inline /admin/* handlers: the 4
  auto-favorite-targets CRUD/run endpoints, load-config (~410 ln),
  ensure/status session-passkey, get-channel, load-owner,
  get-device-metadata, reboot, suppressed-ghosts list/delete, set-time,
  export-config, import-config, and the generic commands dispatcher
  (~310 ln). `bytesToBase64` and the `AUTO_FAVORITE_DEFAULTS` config moved
  in as module-private — this route group is their sole consumer. Two
  dynamic `await import('./services/channelUrlService.js')` calls inside
  the moved get-channel/get-owner handlers were re-pathed to
  '../services/channelUrlService.js' to match the new file's location one
  directory deeper than server.ts.

- server.ts: mount `apiRouter.use('/admin', adminRoutes)`, delete the 34
  inline handlers, the AUTO_FAVORITE_DEFAULTS block, and the bytesToBase64
  helper (now fully relocated), drop the resulting orphaned imports
  (validateFilterNameRegexOnSave, protobufService, autoDeleteByDistanceService,
  getRoutingErrorName, CONFIG_TYPE_MAP/MODULE_FIELD_BY_ID/DEVICE_FIELD_BY_ID,
  getEffectiveDbNodePosition, requireAdmin), and prune the now-orphaned
  handler-specific comments left dangling by the moves. server.ts:
  4,943 -> 2,714 lines.

- The `/auto-ping/stop/:nodeNum` straggler (physically interleaved in the
  same line range) intentionally stays inline — it belongs to nodesRoutes.ts
  in PR3, not this settings/admin PR.

New tests: routes/adminRoutes.test.ts (harness-based) covers the
requireAdmin() gate (401 anon / 403 non-admin / 200 admin on
GET /admin/suppressed-ghosts) plus a representative body-handling
endpoint (PUT /admin/auto-favorite-targets/:nodeNum: write-gate 403,
missing-sourceId 400, and a save-then-read round trip). settingsRoutes.ts
coverage was already real (settingsRoutes.test.ts +
server.settings-persistence.test.ts) and stays green unchanged.

Full Vitest suite: success, 0 failed suites, 0 failed tests
(10,682 total). typecheck: 0 errors. typecheck:tests: 303 errors
(current main parity).
